A lot of dirty, quick fixes. These MUST be changed eventually! It seems to me that Vanilla (which from now on I shall be calling the "Original" Desmume) has an over-reliance on Windows-specific code.
The Wifi markup is just me being very, very angry. Wifi is embedded all over the place, and I cannot even find some of the files that are required. Even Vanilla can't find them. Not sure what they were planning.

+++ source/src/rasterize.cpp
@@ -943,7 +943,12 @@ static bool rasterizerUnitTasksInited = false;
 
 static void* execRasterizerUnit(void* arg)
 {
-	intptr_t which = (intptr_t)arg;
+	//--DCN
+	// After looking at the "vanilla" Desmume... 
+	// [Defined in windows/ram_search.cpp]
+	// I think this is actually an integer:
+	int which = (int)arg;
+	//intptr_t which = (intptr_t)arg;
 	rasterizerUnit[which].mainLoop<true>();
 	return 0;
 }
